+++ internal/provider/models.go
@@ -19,3 +19,11 @@ type YQ struct {
 	Result     types.String `tfsdk:"result"`
 	ID         types.String `tfsdk:"id"`
 }
+
+type GoPluginV1 struct {
+	Plugin    types.String            `tfsdk:"plugin"`
+	InputData types.String            `tfsdk:"input_data"`
+	Vars      map[string]types.String `tfsdk:"vars"`
+	Result    types.String            `tfsdk:"result"`
+	ID        types.String            `tfsdk:"id"`
+}